One ImageSelect to rule them all
