+++ packages/manager/src/features/Linodes/LinodesDetail/LinodeSettings/ImageAndPassword.test.tsx
@@ -14,6 +14,7 @@ const props = {
   onImageChange: vi.fn(),
   onPasswordChange: vi.fn(),
   password: '',
+  selectedImage: '',
   setAuthorizedUsers: vi.fn(),
 };
