+++ packages/manager/src/features/Linodes/LinodesDetail/LinodeSettings/ImageAndPassword.tsx
@@ -2,10 +2,8 @@ import { styled } from '@mui/material/styles';
 import * as React from 'react';
 
 import { AccessPanel } from 'src/components/AccessPanel/AccessPanel';
-import { TargetImageSelect } from 'src/components/TargetImageSelect/TargetImageSelect';
-import { useAllImagesQuery } from 'src/queries/images';
+import { ImageSelect } from 'src/components/ImageSelect/ImageSelect';
 import { useGrants, useProfile } from 'src/queries/profile/profile';
-import { getAPIErrorOrDefault } from 'src/utilities/errorUtils';
 
 import { LinodePermissionsError } from '../LinodePermissionsError';
 
@@ -19,6 +17,7 @@ interface Props {
   onPasswordChange: (password: string) => void;
   password: string;
   passwordError?: string;
+  selectedImage: Image['id'];
   setAuthorizedUsers: (usernames: string[]) => void;
 }
 
@@ -31,17 +30,12 @@ export const ImageAndPassword = (props: Props) => {
     onPasswordChange,
     password,
     passwordError,
+    selectedImage,
     setAuthorizedUsers,
   } = props;
 
   const { data: grants } = useGrants();
   const { data: profile } = useProfile();
-
-  const { data: imagesData, error: imagesError } = useAllImagesQuery();
-
-  const _imagesError = imagesError
-    ? getAPIErrorOrDefault(imagesError, 'Unable to load Images')[0]?.reason
-    : undefined;
 
   const disabled =
     profile?.restricted &&
@@ -50,11 +44,11 @@ export const ImageAndPassword = (props: Props) => {
   return (
     <React.Fragment>
       {disabled && <LinodePermissionsError />}
-      <TargetImageSelect
+      <ImageSelect
         disabled={disabled}
-        errorText={imageFieldError ?? _imagesError}
-        images={imagesData ?? []}
-        onSelect={onImageChange}
+        errorText={imageFieldError}
+        onChange={onImageChange}
+        value={selectedImage}
       />
       <StyledAccessPanel
         disabledReason={
