+++ packages/manager/src/features/Linodes/LinodesDetail/LinodeStorage/CreateDiskDrawer.tsx
@@ -188,6 +188,7 @@ export const CreateDiskDrawer = (props: Props) => {
             authorizedUsers={formik.values.authorized_users}
             linodeId={linodeId}
             password={formik.values.root_pass}
+            selectedImage={formik.values.image}
           />
         )}
         <TextField
